requirements.txt
blinker==1.9.0
cachelib==0.13.0
click==8.3.0
colorama==0.4.6
Flask==3.1.2
Flask-Session==0.8.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
msgspec==0.20.0
mysql-connector-python==9.5.0
python-dotenv==1.2.1
Werkzeug==3.1.3
